feat: triage extracted action items instead of listing them twice
