+++ Logue/App/AppConstants.swift
@@ -26,6 +26,7 @@ enum AppConstants {
         static let documentSortOrder = "documentSortOrder"
         static let meetingSortOrder = "meetingSortOrder"
         static let actionItemSortOrder = "actionItemSortOrder"
+        static let actionItemInboxMode = "actionItemInboxMode"
         /// How the Tasks list is sorted, remembered between launches.
         static let taskSortOrder = "taskSortOrder"
         /// Which filter the Tasks list opens on.
