+++ Logue/Views/ActionItems/ActionItemDashboardView.swift
@@ -1,28 +1,6 @@
 import SwiftUI
 
-// MARK: - Filter & Sort Options
-
-enum ActionItemFilterMode: String, CaseIterable {
-    case all = "All"
-    case pending = "Pending"
-    case overdue = "Overdue"
-    case dueToday = "Due Today"
-    case dueThisWeek = "This Week"
-    case noDueDate = "No Due Date"
-    case completed = "Completed"
-
-    var icon: String {
-        switch self {
-        case .all: "tray.full"
-        case .pending: "circle"
-        case .overdue: "exclamationmark.triangle"
-        case .dueToday: "calendar"
-        case .dueThisWeek: "calendar.badge.clock"
-        case .noDueDate: "calendar.badge.exclamationmark"
-        case .completed: "checkmark.circle"
-        }
-    }
-}
+// MARK: - Sort Options
 
 enum ActionItemSortOrder: String, CaseIterable {
     case dueDateAsc = "Due Date (Soonest)"
@@ -103,17 +81,24 @@ extension DashboardActionItem {
 
 // MARK: - ActionItemDashboardView
 
-/// Cross-meeting dashboard aggregating all action items with filter, sort, and search.
-/// Backend already exists via `MeetingStore` mutations; this surfaces items in one place
-/// so users can triage accountability without visiting each meeting.
+/// The triage inbox for action items the model pulled out of meetings.
+///
+/// Not a second task list. Everything here is awaiting one decision — promote it into Tasks
+/// or dismiss it — and anything decided leaves. The due-date vocabulary this screen used to
+/// carry now lives on the Tasks surface, which is where work the user accepted belongs.
 struct ActionItemDashboardView: View {
     @Environment(MeetingStore.self) private var meetingStore
     @State private var taskStore = TaskStore.shared
 
     @State private var searchText = ""
-    @State private var filterMode: ActionItemFilterMode = .pending
+    @AppStorage(AppConstants.UserDefaultsKeys.actionItemInboxMode)
+    private var inboxModeRaw = ActionItemInboxMode.inbox.rawValue
     @AppStorage(AppConstants.UserDefaultsKeys.actionItemSortOrder)
     private var sortOrderRaw = ActionItemSortOrder.dueDateAsc.rawValue
+
+    private var inboxMode: ActionItemInboxMode {
+        ActionItemInboxMode(rawValue: inboxModeRaw) ?? .inbox
+    }
 
     private var sortOrder: ActionItemSortOrder {
         ActionItemSortOrder(rawValue: sortOrderRaw) ?? .dueDateAsc
@@ -143,22 +128,11 @@ struct ActionItemDashboardView: View {
     }
 
     private func matchesFilter(_ item: DashboardActionItem) -> Bool {
-        switch filterMode {
-        case .all:
-            true
-        case .pending:
-            !item.actionItem.isCompleted
-        case .overdue:
-            item.urgency == .overdue
-        case .dueToday:
-            item.urgency == .dueToday
-        case .dueThisWeek:
-            item.urgency == .dueThisWeek || item.urgency == .dueToday || item.urgency == .overdue
-        case .noDueDate:
-            !item.actionItem.isCompleted && item.actionItem.dueDate == nil
-        case .completed:
-            item.actionItem.isCompleted
-        }
+        ActionItemInbox.matches(
+            item.actionItem,
+            mode: inboxMode,
+            isPromoted: taskStore.promotedTask(for: item.actionItem.id) != nil
+        )
     }
 
     private func matchesSearch(_ item: DashboardActionItem) -> Bool {
@@ -201,21 +175,10 @@ struct ActionItemDashboardView: View {
 
     // MARK: - Counts for filter chips
 
-    private var counts: [ActionItemFilterMode: Int] {
-        var dict: [ActionItemFilterMode: Int] = [:]
-        let items = allItems
-        dict[.all] = items.count
-        dict[.pending] = items.filter { !$0.actionItem.isCompleted }.count
-        dict[.overdue] = items.filter { $0.urgency == .overdue }.count
-        dict[.dueToday] = items.filter { $0.urgency == .dueToday }.count
-        dict[.dueThisWeek] = items.filter {
-            $0.urgency == .dueToday || $0.urgency == .dueThisWeek || $0.urgency == .overdue
-        }.count
-        dict[.noDueDate] = items.filter {
-            !$0.actionItem.isCompleted && $0.actionItem.dueDate == nil
-        }.count
-        dict[.completed] = items.filter(\.actionItem.isCompleted).count
-        return dict
+    private var counts: [ActionItemInboxMode: Int] {
+        ActionItemInbox.counts(allItems.map(\.actionItem)) { item in
+            taskStore.promotedTask(for: item.id) != nil
+        }
     }
 
     // MARK: - Body
@@ -242,6 +205,9 @@ struct ActionItemDashboardView: View {
 
     private var subtitle: String {
         let total = filteredItems.count
+        if inboxMode == .inbox {
+            return "\(total) in inbox"
+        }
         return "\(total) item\(total == 1 ? "" : "s")"
     }
 
@@ -297,16 +263,16 @@ struct ActionItemDashboardView: View {
     private var filterChipBar: some View {
         ScrollView(.horizontal, showsIndicators: false) {
             HStack(spacing: 6) {
-                ForEach(ActionItemFilterMode.allCases, id: \.rawValue) { mode in
+                ForEach(ActionItemInboxMode.allCases, id: \.rawValue) { mode in
                     let count = counts[mode] ?? 0
                     FilterChip(
-                        label: "\(mode.rawValue) \(count)",
-                        isSelected: filterMode == mode,
+                        label: "\(mode.displayName) \(count)",
+                        isSelected: inboxMode == mode,
                         tintColor: tintColor(for: mode)
                     ) {
-                        filterMode = mode
+                        inboxModeRaw = mode.rawValue
                     }
-                    .accessibilityLabel("\(mode.rawValue), \(count) items")
+                    .accessibilityLabel("\(mode.displayName), \(count) items")
                 }
             }
             .padding(.horizontal, 24)
@@ -314,11 +280,9 @@ struct ActionItemDashboardView: View {
         }
     }
 
-    private func tintColor(for mode: ActionItemFilterMode) -> Color? {
+    private func tintColor(for mode: ActionItemInboxMode) -> Color? {
         switch mode {
-        case .overdue: AppThemeConstants.error
-        case .dueToday, .dueThisWeek: AppThemeConstants.warning
-        case .completed: AppThemeConstants.success
+        case .dismissed: AppThemeConstants.warning
         default: nil
         }
     }
@@ -329,7 +293,7 @@ struct ActionItemDashboardView: View {
         ScrollView {
             LazyVStack(spacing: 0) {
                 ForEach(filteredItems) { item in
-                    ActionItemDashboardRow(item: item)
+                    ActionItemInboxRow(item: item)
                     if item.id != filteredItems.last?.id {
                         Divider().padding(.leading, 44)
                     }
@@ -355,14 +319,10 @@ struct ActionItemDashboardView: View {
         if !searchText.isEmpty {
             return "No Matching Items"
         }
-        switch filterMode {
+        switch inboxMode {
+        case .inbox: return "Inbox Zero"
+        case .dismissed: return "Nothing Dismissed"
         case .all: return "No Action Items Yet"
-        case .pending: return "All Caught Up"
-        case .overdue: return "Nothing Overdue"
-        case .dueToday: return "Nothing Due Today"
-        case .dueThisWeek: return "Nothing Due This Week"
-        case .noDueDate: return "All Items Have Due Dates"
-        case .completed: return "No Completed Items"
         }
     }
 
@@ -370,219 +330,20 @@ struct ActionItemDashboardView: View {
         if !searchText.isEmpty {
             return "magnifyingglass"
         }
-        switch filterMode {
-        case .overdue, .dueToday, .dueThisWeek: return "checkmark.circle"
-        case .completed: return "circle"
-        default: return "checklist"
-        }
+        return inboxMode == .inbox ? "tray" : "checklist"
     }
 
     private var emptyDescription: String {
         if !searchText.isEmpty {
             return "No action items match \"\(searchText)\""
         }
-        switch filterMode {
-        case .all: return "Action items extracted from meetings will appear here."
-        case .pending: return "Every action item is either completed or scheduled. Nice work."
-        case .overdue: return "No action items are past their due date."
-        case .dueToday: return "No action items are due today."
-        case .dueThisWeek: return "No action items are due in the next seven days."
-        case .noDueDate: return "Every pending item has a due date assigned."
-        case .completed: return "Completed items will appear here once you mark them done."
+        switch inboxMode {
+        case .inbox:
+            return "Every action item Logue found has been added to your tasks or dismissed."
+        case .dismissed:
+            return "Items you decide aren't worth acting on will appear here."
+        case .all:
+            return "Action items extracted from meetings will appear here."
         }
-    }
-}
-
-// MARK: - Dashboard Row
-
-/// A single row in the Action Item Dashboard.
-private struct ActionItemDashboardRow: View {
-    let item: DashboardActionItem
-    @Environment(MeetingStore.self) private var meetingStore
-    @State private var taskStore = TaskStore.shared
-    @State private var isHovered = false
-
-    /// Reflects the promoted state rather than offering the same action twice.
-    ///
-    /// Promotion is idempotent, so pressing it again would be harmless — but a button that
-    /// looks like it has never been pressed reads as "that did nothing".
-    @ViewBuilder
-    private var promoteControl: some View {
-        if taskStore.promotedTask(for: item.actionItem.id) == nil {
-            Button {
-                taskStore.promote(item.actionItem, from: item.meetingID)
-            } label: {
-                Image(systemName: "arrow.right.circle")
-            }
-            .buttonStyle(.plain)
-            .foregroundStyle(.secondary)
-            .help("Add this to your tasks")
-            .accessibilityLabel("Add to Tasks")
-            // Shown on hover only: one of these per row, permanently visible, would compete
-            // with the due badge for the eye.
-            .opacity(isHovered ? 1 : 0)
-        } else {
-            Image(systemName: "checkmark.circle.fill")
-                .font(.caption)
-                .foregroundStyle(.tertiary)
-                .help("Already on your task list")
-                .accessibilityLabel("Already in Tasks")
-        }
-    }
-
-    var body: some View {
-        HStack(alignment: .center, spacing: 10) {
-            completionButton
-
-            VStack(alignment: .leading, spacing: 3) {
-                Text(item.actionItem.title)
-                    .font(.callout.weight(.medium))
-                    .foregroundStyle(item.actionItem.isCompleted ? .secondary : .primary)
-                    .strikethrough(item.actionItem.isCompleted)
-                    .lineLimit(2)
-
-                HStack(spacing: 8) {
-                    meetingLink
-
-                    if let assignee = item.actionItem.assignee, !assignee.isEmpty {
-                        Text("·")
-                            .font(.caption)
-                            .foregroundStyle(.tertiary)
-                        HStack(spacing: 3) {
-                            Image(systemName: "person")
-                                .font(.caption2)
-                            Text(assignee)
-                                .font(.caption)
-                        }
-                        .foregroundStyle(.secondary)
-                    }
-                }
-            }
-
-            Spacer()
-
-            promoteControl
-
-            if let dueDate = item.actionItem.dueDate {
-                dueBadge(dueDate)
-            } else if !item.actionItem.isCompleted {
-                Text("No due date")
-                    .font(.caption)
-                    .foregroundStyle(.tertiary)
-            }
-        }
-        .padding(.horizontal, 8)
-        .padding(.vertical, 10)
-        .background(
-            isHovered
-                ? AppThemeConstants.surfaceBackground
-                : Color.clear,
-            in: RoundedRectangle(cornerRadius: AppThemeConstants.radiusSmall)
-        )
-        .contentShape(Rectangle())
-        .onHover { isHovered = $0 }
-        .onTapGesture {
-            openMeeting()
-        }
-        .contextMenu {
-            Button {
-                openMeeting()
-            } label: {
-                Label("Open Meeting", systemImage: "arrow.up.right.square")
-            }
-            Button {
-                meetingStore.toggleActionItemCompleted(
-                    itemID: item.actionItem.id,
-                    in: item.meetingID
-                )
-            } label: {
-                Label(
-                    item.actionItem.isCompleted ? "Mark Incomplete" : "Mark Complete",
-                    systemImage: item.actionItem.isCompleted ? "circle" : "checkmark.circle"
-                )
-            }
-        }
-        // Accessibility: `.combine` was flattening + stripping the explicit label in macOS 26
-        // SwiftUI. `.contain` keeps inner buttons individually reachable (completion toggle,
-        // meeting link); the explicit label + isButton trait + primary action expose the row
-        // itself as a named, activatable AXButton.
-        .accessibilityElement(children: .contain)
-        .accessibilityLabel(accessibilityText)
-        .accessibilityAddTraits(.isButton)
-        .accessibilityAction {
-            openMeeting()
-        }
-    }
-
-    private var completionButton: some View {
-        Button {
-            HapticFeedback.levelChange()
-            meetingStore.toggleActionItemCompleted(
-                itemID: item.actionItem.id,
-                in: item.meetingID
-            )
-        } label: {
-            Image(systemName: item.actionItem.isCompleted ? "checkmark.circle.fill" : "circle")
-                .font(.title3)
-                .foregroundStyle(
-                    item.actionItem.isCompleted
-                        ? AppThemeConstants.success
-                        : .secondary
-                )
-        }
-        .buttonStyle(.plain)
-        .accessibilityLabel(item.actionItem.isCompleted ? "Mark incomplete" : "Mark complete")
-    }
-
-    private var meetingLink: some View {
-        Button {
-            openMeeting()
-        } label: {
-            HStack(spacing: 3) {
-                Image(systemName: "waveform")
-                    .font(.caption2)
-                Text(item.meetingTitle)
-                    .font(.caption)
-                    .lineLimit(1)
-            }
-            .foregroundStyle(AppThemeConstants.accent)
-        }
-        .buttonStyle(.plain)
-    }
-
-    private func dueBadge(_ date: Date) -> some View {
-        let color = item.urgency.color
-        return HStack(spacing: 4) {
-            Image(systemName: "calendar")
-                .font(.caption2)
-            Text(date.formatted(.relative(presentation: .named)))
-                .font(.caption.weight(.medium))
-        }
-        .foregroundStyle(color)
-        .padding(.horizontal, 8)
-        .padding(.vertical, 4)
-        .background(
-            color.opacity(AppThemeConstants.opacityLight),
-            in: RoundedRectangle(cornerRadius: 6)
-        )
-    }
-
-    private func openMeeting() {
-        meetingStore.selectedMeetingID = item.meetingID
-    }
-
-    private var accessibilityText: String {
-        var parts: [String] = [item.actionItem.title]
-        if item.actionItem.isCompleted {
-            parts.append("completed")
-        }
-        parts.append("from \(item.meetingTitle)")
-        if let due = item.actionItem.dueDate {
-            parts.append("due \(due.formatted(.relative(presentation: .named)))")
-        }
-        if let assignee = item.actionItem.assignee, !assignee.isEmpty {
-            parts.append("assigned to \(assignee)")
-        }
-        return parts.joined(separator: ", ")
     }
 }
